http server accepts queries

+++ src/main/scala/graph/package.scala
@@ -1,5 +1,5 @@
 package object graph {
-    type FeatVec = Seq[Float]
+    type FeatVec = Seq[Double]
     type FeatMat = Seq[FeatVec]
     type Id = Int
